Create new_language_wordlisting.py
This took me an hour to come this far, and yet I have not been able to make the full program. I will try later Insha Allah.

Target is to create a language which converts every word from input by following rules
1> if a word beings with vowel then add 'yay' to it
2> otherwise push all non vowels to end and add 'ay' to end of the word.

What I have accomplished in an hour is created two lists, one with words beginning with vowels and the other with the rest. But this was not the objective. So, I will try again and finish it tomorrow Insha Allah.